+++ changelog.md
@@ -5,6 +5,8 @@
 ### Features
 - [#3476](https://github.com/ignite/cli/pull/3476) Use `buf.build` binary to generate the proto files
 
+- [#3505](https://github.com/ignite/cli/pull/3505) Auto migrate dependency tools
+
 ### Changes
 
 - [#3444](https://github.com/ignite/cli/pull/3444) Add support for ICS chains in ts-client generation